requirements.txt
flask>=3.0
flask-cors>=4.0
Pillow>=10.0
opencv-python>=4.8
roboflow>=1.0
google-api-python-client>=2.0
google-auth-httplib2>=0.2
google-auth-oauthlib>=1.0
ultralytics>=8.0
